ajout des types requis pour les UE mise en place des models pour les UE ajout mock
