+++ app/src/scripts/API/APITypes/UETypes.ts
@@ -1,9 +1,9 @@
 import {Course} from "./Course.ts";
 
 export type UE = {
-    academic_year: number;
-    courses: Course[];
     id: string;
     name: string;
+    courses: Course[];
+    academic_year: number;
 }
